Complete separation of application into components
