+++ controller/app.json
@@ -2,7 +2,7 @@
   "expo": {
     "name": "EZ-RASSOR Controller",
     "description": "This controller application has been made for the EZ-RASSOR, an excavation robot that runs on ROS.",
-    "slug": "EZ-RASSOR Controller",
+    "slug": "EZRASSORController",
     "privacy": "public",
     "sdkVersion": "31.0.0",
     "platforms": ["ios", "android"],
