+++ controller/rn-cli.config.js
@@ -1,0 +1,7 @@
+// This tweak allows for the use of jsx files in react-native
+// More information can be found at https://github.com/facebook/react-native/pull/5233#issuecomment-382083236``
+module.exports = {
+    resolver: {
+        sourceExts: ['jsx', 'js'],
+    },
+};
